requirements.txt
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
frozenlist==1.3.3
greenlet==2.0.2
idna==3.4
multidict==6.0.4
numpy==1.25.0
pandas==2.0.3
psycopg2-binary==2.9.6
PyJWT==2.7.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.18
twilio==8.4.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.3
yarl==1.9.2
